+++ src/__tests__/completion/basic.test.ts
@@ -352,6 +352,19 @@ describe('completion', () => {
       completion.stop(true)
       spy.mockRestore()
     })
+
+    it('should disable filter on backspace', async () => {
+      helper.updateConfiguration('suggest.filterOnBackspace', false)
+      await create(['this', 'thoit'], true)
+      await nvim.input('this')
+      await helper.waitValue(() => {
+        return completion.activeItems.length
+      }, 1)
+      await nvim.input('<bs>')
+      await helper.waitValue(() => {
+        return completion.isActivated
+      }, false)
+    })
   })
 
   describe('suggest variables', () => {
@@ -366,7 +379,7 @@ describe('completion', () => {
       let doc = await workspace.document
       await doc.buffer.setVar('coc_suggest_disable', 1)
       await nvim.input('if')
-      await helper.wait(30)
+      await helper.wait(20)
       let visible = await pumvisible()
       expect(visible).toBe(false)
     })
@@ -375,7 +388,7 @@ describe('completion', () => {
       let doc = await workspace.document
       await doc.buffer.setVar('coc_disabled_sources', ['foo'])
       await nvim.input('if')
-      await helper.wait(30)
+      await helper.wait(20)
       let visible = await pumvisible()
       expect(visible).toBe(false)
     })
@@ -730,7 +743,6 @@ describe('completion', () => {
     })
 
     it('should not complete inComplete source when isIncomplete is false', async () => {
-      let lastOption: CompleteOption
       let source: ISource = {
         priority: 0,
         enable: true,
@@ -738,7 +750,6 @@ describe('completion', () => {
         sourceType: SourceType.Service,
         triggerCharacters: ['.'],
         doComplete: async (opt: CompleteOption) => {
-          lastOption = opt
           await helper.wait(30)
           if (opt.input.length <= 1) {
             return { isIncomplete: true, items: [{ word: 'foobar' }] }
@@ -809,6 +820,16 @@ describe('completion', () => {
       await helper.wait(10)
       await nvim.input('.')
       await helper.waitFor('getline', ['.'], 'foo.')
+    })
+
+    it('should filter on backspace', async () => {
+      await create(['foo', 'fbi'], false)
+      await nvim.input('f')
+      await helper.waitPopup()
+      await nvim.input('o')
+      await helper.waitValue(() => completion.activeItems.length, 1)
+      await nvim.input('<backspace>')
+      await helper.waitValue(() => completion.activeItems.length, 2)
     })
   })