+++ history.md
@@ -2,6 +2,7 @@
 
 - Make `vscode.open` command work with file uri.
 - Cancel option for `workspace.registerExprKeymap()`.
+- Support `suggest.filterOnBackspace` configuration.
 
 # 2023-01-22